code:change double on int

+++ main.cpp
@@ -4,7 +4,7 @@ using namespace std;
 int main()
 {
     cout << "Enter A and B: ";
-    double a, b;
+    int a, b;
     cin >> a >> b;
     cout << "A + B = " << a + b << '\n'
      << "A - B = " << a - b << '\n'
